+++ CHANGELOG.md
@@ -4,3 +4,10 @@ All Notable changes to `kudaapitken` will be documented in this file
 
 ## 2022-07-21
 - Initial release
+
+## 2022-12-23
+- updated virtual account services
+- updated service types
+= Added Biller
+- Added Saving
+- Added Giftcard